fix: move backstab prediction logic to shared
Agent-Logs-Url: https://github.com/Fooksy2304/sunrise-station-med/sessions/098b2a85-5af8-4b1d-af20-093c28a1f08a
